wip refactoring

+++ src/Interface/WatchdogInterface.h
@@ -14,16 +14,16 @@ namespace Mengine
         SERVICE_DECLARE( "Watchdog" )
 
     public:
-        virtual float watch( const String & _tag ) = 0;
+        virtual double watch( const String & _tag ) = 0;
     };
     //////////////////////////////////////////////////////////////////////////
 #define WATCHDOG_SERVICE()\
 	SERVICE_GET(Mengine::WatchdogInterface)
     //////////////////////////////////////////////////////////////////////////
 #ifdef MENGINE_MASTER_RELEASE
-#	define WATCHDOG( tag ) (0.f)
+#	define WATCHDOG( tag ) (0.0)
 #else
 #	define WATCHDOG( tag )\
-	(SERVICE_EXIST(Mengine::WatchdogInterface) ? WATCHDOG_SERVICE()->watch(tag) : 0.f)
+	(SERVICE_EXIST(Mengine::WatchdogInterface) ? WATCHDOG_SERVICE()->watch(tag) : 0.0)
 #endif
 }
